code/requirements.txt
CoffeeScript==1.0.10
Django==1.7.1
Pillow==2.6.1
PyExecJS==1.0.5
Werkzeug==0.9.6
astroid==1.2.1
coverage==3.7.1
django-bootstrap3==4.11.0
django-bootstrap3-datetimepicker==2.2.3
django-datetime-widget==0.9.2
django-extensions==1.4.6
django-jenkins==0.16.3
django-multiselectfield==0.1.3
django-registration-redux==1.1
easy-thumbnails==2.2
extypes==0.1.0
logilab-common==0.63.0
pep8==1.5.7
pyflakes==0.8.1
pylint==1.3.1
pytz==2014.9
six==1.8.0
